+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 [Unreleased](https://github.com/jshwi/hin/compare/v0.2.0...HEAD)
 ------------------------------------------------------------------------
+### Fixed
+- print when there are no changes to push to remote
 
 [0.2.0](https://github.com/jshwi/hin/releases/tag/v0.2.0) - 2023-06-17
 ------------------------------------------------------------------------